Retry HostInfo startup and centralize host detection
Previously, if HostInfo failed during startup, it caused the controller manager to stop
while the daemon still reported the control plane as ready. Since no reconcilers were
running, nothing could recover it. Instead, retry startup until it succeeds or the daemon
shuts down.

Also, move host detection into a shared pkg/hostinfo package so both the webhooks and HostInfo
reconciler use the same Detect() logic. This ensures they always report and validate against
the same host resource limits.
